{Update} Player / Collision Scripts
CollisionScript: instead of directly destroying the knight upon contact with an obstacle, now calls function in PlayerScript, can now enter in contact with Bonuses if they have the "Bonus" tag.
PlayerScript: Various functions for when a knight gets a bonus / run into an obstacle. Not completed yet (it's more like a skeleton)